requirements.txt
streamlit>=1.37.0
openai>=1.50.0
groq>=0.9.0
requests>=2.32.0
python-dotenv>=1.0.1
sentence-transformers>=3.0.0
chromadb>=0.5.0
pdfplumber>=0.11.0
python-docx>=1.1.0
whisper>=1.1.10
pyttsx3>=2.90
soundfile>=0.12.1
